finished integral open address hashset tests. Began work on float, double, and ref open address hashset tests

+++ Omni-Impl/test-new/omni/impl/DataType.java
@@ -1273,7 +1273,7 @@ public enum DataType{
         }
         @Override
         public Object convertVal(float val){
-            return (double)val;
+            return val;
         }
         @Override
         public Object convertVal(double val){
